requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
Django==4.2.4
djangorestframework==3.14.0
drf-spectacular==0.26.4
importlib-resources==6.0.1
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
zipp==3.16.2
